ENG-01: extract Resampler; make methods the aggregator and _pca_pls a shim
Move Resampler into _resampling. methods.py now imports the public surface
directly from the focused submodules (the canonical re-exporter), and
_pca_pls is reduced to a thin backward-compat shim that re-exports from
methods plus the private _nz helper. _pca_pls.py drops from ~6,200 lines to
~25.
